move LogoChain to separate file
